requirements.txt
sentence-transformers
faiss-cpu
streamlit
tqdm
python-dotenv
numpy
pandas